Split out CoverageDetails
