+++ lib/react/coveragedetails/countdetails.tsx
@@ -1,0 +1,38 @@
+import {
+    LineChart, //
+    Line,
+    XAxis,
+    YAxis,
+    CartesianGrid,
+    Tooltip,
+    Legend,
+    ResponsiveContainer
+} from 'recharts';
+
+import {findIndex as _findIndex, reduce as _reduce, debounce as _debounce, map as _map} from 'lodash';
+
+import graphcolours from '../graphcolours';
+
+export function CountDetails(props: {byDay: any; c: number}) {
+    return (
+        <>
+            Number of packets: {props.c}
+            <br />
+            {props.byDay && props.byDay.length > 0 ? (
+                <>
+                    <br />
+                    <ResponsiveContainer width="100%" height={150}>
+                        <LineChart width={480} height={180} data={props.byDay} margin={{top: 5, right: 30, left: 20, bottom: 5}} syncId="date">
+                            <CartesianGrid strokeDasharray="3 3" />
+                            <XAxis dataKey="date" style={{fontSize: '0.8rem'}} />
+                            <YAxis style={{fontSize: '0.8rem'}} />
+                            <Tooltip />
+                            <Legend />
+                            <Line name="Count" isAnimationActive={false} type="monotone" dataKey="count" stroke={graphcolours[0]} dot={{r: 1}} />
+                        </LineChart>
+                    </ResponsiveContainer>
+                </>
+            ) : null}
+        </>
+    );
+}
