+++ lib/react/coveragedetails/waitforgraph.tsx
@@ -1,0 +1,11 @@
+import {FaUserClock, FaChartBar} from 'react-icons/fa';
+
+export function WaitForGraph() {
+    return (
+        <>
+            <FaChartBar />
+            &nbsp;
+            <FaUserClock />
+        </>
+    );
+}
